requirements.txt
django
djangorestframework
drf-yasg
djangorestframework-simplejwt
gunicorn
python-decouple
psycopg2-binary
reportlab   
openai==1.12.0
httpx==0.27.0
celery
redis
django-celery-results
django-redis
langchain==0.1.16
langgraph==0.0.27
tiktoken
faiss-cpu
pydantic==1.10.13
langchain-openai==0.0.5
langchain-community==0.0.33